+++ packages/gatsby-plugin-image/src/components/hooks.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-expressions */
 import {
   useState,
   CSSProperties,
@@ -17,6 +18,9 @@ import {
   IGatsbyImageHelperArgs,
   generateImageData,
   Layout,
+  EVERY_BREAKPOINT,
+  IImage,
+  ImageFormat,
 } from "../image-utils"
 const imageCache = new Set<string>()
 
@@ -54,20 +58,22 @@ export function getWrapperProps(
 ): Pick<HTMLAttributes<HTMLElement>, "className" | "style"> & {
   "data-gatsby-image-wrapper": string
 } {
-  const wrapperStyle: CSSProperties = {
-    position: `relative`,
-    overflow: `hidden`,
-  }
+  const wrapperStyle: CSSProperties = {}
+
+  let className = `gatsby-image-wrapper`
 
   if (layout === `fixed`) {
     wrapperStyle.width = width
     wrapperStyle.height = height
   } else if (layout === `constrained`) {
-    wrapperStyle.display = `inline-block`
+    if (!global.GATSBY___IMAGE) {
+      wrapperStyle.display = `inline-block`
+    }
+    className = `gatsby-image-wrapper gatsby-image-wrapper-constrained`
   }
 
   return {
-    className: `gatsby-image-wrapper`,
+    className,
     "data-gatsby-image-wrapper": ``,
     style: wrapperStyle,
   }
@@ -84,24 +90,114 @@ export async function applyPolyfill(
   ;(window as any).objectFitPolyfill(ref.current)
 }
 
-export function useGatsbyImage({
+export interface IUrlBuilderArgs<OptionsType> {
+  width: number
+  height: number
+  baseUrl: string
+  format: ImageFormat
+  options: OptionsType
+}
+export interface IGetImageDataArgs<OptionsType = {}> {
+  baseUrl: string
+  /**
+   * For constrained and fixed images, the size of the image element
+   */
+  width?: number
+  height?: number
+  /**
+   * If available, pass the source image width and height
+   */
+  sourceWidth?: number
+  sourceHeight?: number
+  /**
+   * If only one dimension is passed, then this will be used to calculate the other.
+   */
+  aspectRatio?: number
+  layout?: Layout
+  /**
+   * Returns a URL based on the passed arguments. Should be a pure function
+   */
+  urlBuilder: (args: IUrlBuilderArgs<OptionsType>) => string
+
+  /**
+   * Should be a data URI
+   */
+  placeholderURL?: string
+  backgroundColor?: string
+  /**
+   * Used in error messages etc
+   */
+  pluginName?: string
+
+  /**
+   * If you do not support auto-format, pass an array of image types here
+   */
+  formats?: Array<ImageFormat>
+
+  breakpoints?: Array<number>
+
+  /**
+   * Passed to the urlBuilder function
+   */
+  options?: OptionsType
+}
+
+/**
+ * Use this hook to generate gatsby-plugin-image data in the browser.
+ */
+export function getImageData<OptionsType>({
+  baseUrl,
+  urlBuilder,
+  sourceWidth,
+  sourceHeight,
   pluginName = `useGatsbyImage`,
-  ...args
-}: IGatsbyImageHelperArgs): IGatsbyImageData {
-  // TODO: use context to get default plugin options and spread them in here
-  return generateImageData({ pluginName, ...args })
+  formats = [`auto`],
+  breakpoints = EVERY_BREAKPOINT,
+  options,
+  ...props
+}: IGetImageDataArgs<OptionsType>): IGatsbyImageData {
+  const generateImageSource = (
+    baseUrl: string,
+    width: number,
+    height?: number,
+    format?: ImageFormat
+  ): IImage => {
+    return {
+      width,
+      height,
+      format,
+      src: urlBuilder({ baseUrl, width, height, options, format }),
+    }
+  }
+
+  const sourceMetadata: IGatsbyImageHelperArgs["sourceMetadata"] = {
+    width: sourceWidth,
+    height: sourceHeight,
+    format: `auto`,
+  }
+
+  const args: IGatsbyImageHelperArgs = {
+    ...props,
+    pluginName,
+    generateImageSource,
+    filename: baseUrl,
+    formats,
+    breakpoints,
+    sourceMetadata,
+  }
+  return generateImageData(args)
 }
 
 export function getMainProps(
   isLoading: boolean,
   isLoaded: boolean,
-  images: any,
+  images: IGatsbyImageData["images"],
   loading?: "eager" | "lazy",
   toggleLoaded?: (loaded: boolean) => void,
   cacheKey?: string,
   ref?: RefObject<HTMLImageElement>,
   style: CSSProperties = {}
-): MainImageProps {
+): Partial<MainImageProps> {
   const onLoad: ReactEventHandler<HTMLImageElement> = function (e) {
     if (isLoaded) {
       return
@@ -279,4 +375,82 @@ export function useImageLoaded(
     isLoaded,
     toggleLoaded,
   }
+}
+
+export interface IArtDirectedImage {
+  media: string
+  image: IGatsbyImageData
+}
+
+/**
+ * Generate a Gatsby image data object with multiple, art-directed images that display at different
+ * resolutions.
+ *
+ * @param defaultImage The image displayed when no media query matches.
+ * It is also used for all other settings applied to the image, such as width, height and layout.
+ * You should pass a className to the component with media queries to adjust the size of the container,
+ * as this cannot be adjusted automatically.
+ * @param artDirected Array of objects which each contains a `media` string which is a media query
+ * such as `(min-width: 320px)`, and the image object to use when that query matches.
+ */
+export function withArtDirection(
+  defaultImage: IGatsbyImageData,
+  artDirected: Array<IArtDirectedImage>
+): IGatsbyImageData {
+  const { images, placeholder, ...props } = defaultImage
+  const output: IGatsbyImageData = {
+    ...props,
+    images: {
+      ...images,
+      sources: [],
+    },
+    placeholder: placeholder && {
+      ...placeholder,
+      sources: [],
+    },
+  }
+
+  artDirected.forEach(({ media, image }) => {
+    if (!media) {
+      if (process.env.NODE_ENV === `development`) {
+        console.warn(
+          "[gatsby-plugin-image] All art-directed images passed to must have a value set for `media`. Skipping."
+        )
+      }
+      return
+    }
+
+    if (
+      image.layout !== defaultImage.layout &&
+      process.env.NODE_ENV === `development`
+    ) {
+      console.warn(
+        `[gatsby-plugin-image] Mismatched image layout: expected "${defaultImage.layout}" but received "${image.layout}". All art-directed images use the same layout as the default image`
+      )
+    }
+
+    output.images.sources.push(
+      ...image.images.sources.map(source => {
+        return { ...source, media }
+      }),
+      {
+        media,
+        srcSet: image.images.fallback.srcSet,
+      }
+    )
+
+    if (!output.placeholder) {
+      return
+    }
+
+    output.placeholder.sources.push({
+      media,
+      srcSet: image.placeholder.fallback,
+    })
+  })
+  output.images.sources.push(...images.sources)
+  if (placeholder?.sources) {
+    output.placeholder?.sources.push(...placeholder.sources)
+  }
+  return output
 }