+++ packages/gatsby-plugin-image/src/index.ts
@@ -6,7 +6,15 @@ export {
 export { Placeholder } from "./components/placeholder"
 export { MainImage } from "./components/main-image"
 export { StaticImage } from "./components/static-image.server"
-export { getImage, getSrc, useGatsbyImage } from "./components/hooks"
+export {
+  getImage,
+  getSrc,
+  getImageData,
+  withArtDirection,
+  IArtDirectedImage,
+  IGetImageDataArgs,
+  IUrlBuilderArgs,
+} from "./components/hooks"
 export {
   generateImageData,
   IGatsbyImageHelperArgs,