+++ packages/gatsby-source-wordpress/.babelrc.js
@@ -1,0 +1,42 @@
+module.exports = {
+  presets: [[`babel-preset-gatsby-package`]],
+  plugins: [
+    [
+      `@babel/plugin-proposal-private-methods`,
+      {
+        loose: true,
+      },
+    ],
+    [
+      `babel-plugin-module-resolver`,
+      {
+        alias: {
+          "~": `./src`,
+        },
+      },
+    ],
+    [
+      `@babel/plugin-proposal-class-properties`,
+      {
+        loose: true,
+      },
+    ],
+    [
+      `import-globals`,
+      {
+        dd: {
+          moduleName: `dumper.js`,
+          exportName: `dd`,
+        },
+        dump: {
+          moduleName: `dumper.js`,
+          exportName: `dump`,
+        },
+        clipboardy: {
+          moduleName: `clipboardy`,
+          exportName: `default`,
+        },
+      },
+    ],
+  ],
+}