+++ packages/gatsby/src/schema/infer/add-inferred-fields.js
@@ -22,9 +22,6 @@ const addInferredFields = ({
     typeComposer,
     defaults: {
       shouldAddFields: true,
-      shouldAddDefaultResolvers: typeComposer.hasExtension(`infer`)
-        ? false
-        : true,
     },
   })
   addInferredFieldsImpl({
@@ -95,53 +92,6 @@ const addInferredFieldsImpl = ({
       if (config.shouldAddFields) {
         typeComposer.addFields({ [key]: fieldConfig })
         typeComposer.setFieldExtension(key, `createdFrom`, `inference`)
-      }
-    } else {
-      // Deprecated, remove in v3
-      if (config.shouldAddDefaultResolvers) {
-        // Add default resolvers to existing fields if the type matches
-        // and the field has neither args nor resolver explicitly defined.
-        const field = typeComposer.getField(key)
-        if (
-          field.type.toString().replace(/[[\]!]/g, ``) ===
-            fieldConfig.type.toString() &&
-          _.isEmpty(field.args) &&
-          !field.resolve
-        ) {
-          const { extensions } = fieldConfig
-          if (extensions) {
-            Object.keys(extensions)
-              .filter(name =>
-                // It is okay to list allowed extensions explicitly here,
-                // since this is deprecated anyway and won't change.
-                [`dateformat`, `fileByRelativePath`, `link`, `proxy`].includes(
-                  name
-                )
-              )
-              .forEach(name => {
-                if (!typeComposer.hasFieldExtension(key, name)) {
-                  typeComposer.setFieldExtension(key, name, extensions[name])
-
-                  const typeName = typeComposer.getTypeName()
-                  const implementsNode =
-                    unsanitizedFieldPath.length === 1 ? `implements Node ` : ``
-                  const extension = printDirectives(
-                    { [name]: extensions[name] },
-                    schemaComposer.getDirectives()
-                  )
-                  report.warn(
-                    `Deprecation warning: adding inferred extension \`${name}\` for field \`${typeName}.${key}\`.\n` +
-                      `In Gatsby v3, only fields with an explicit directive/extension will be resolved correctly.\n` +
-                      `Add the following type definition to fix this:\n\n` +
-                      `  type ${typeComposer.getTypeName()} ${implementsNode}{\n` +
-                      `    ${key}: ${field.type.toString()}${extension}\n` +
-                      `  }\n\n` +
-                      `https://www.gatsbyjs.com/docs/actions/#createTypes`
-                  )
-                }
-              })
-          }
-        }
       }
     }
   })
@@ -358,9 +308,7 @@ const getSimpleFieldConfig = ({
           if (lists !== arrays) return null
         } else {
           // When the field type has not been explicitly defined, we
-          // don't need to continue in case of @dontInfer, because
-          // "addDefaultResolvers: true" only makes sense for
-          // pre-existing types.
+          // don't need to continue in case of @dontInfer
           if (!config.shouldAddFields) return null
 
           const typeName = createTypeName(selector)
@@ -448,8 +396,5 @@ const getInferenceConfig = ({ typeComposer, defaults }) => {
     shouldAddFields: typeComposer.hasExtension(`infer`)
       ? typeComposer.getExtension(`infer`)
       : defaults.shouldAddFields,
-    shouldAddDefaultResolvers: typeComposer.hasExtension(`addDefaultResolvers`)
-      ? typeComposer.getExtension(`addDefaultResolvers`)
-      : defaults.shouldAddDefaultResolvers,
   }
 }