+++ packages/gatsby-cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@3.0.0-next.1...gatsby-cli@3.0.0-next.2) (2021-02-11)
+
+**Note:** Version bump only for package gatsby-cli
+
+# [3.0.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@3.0.0-next.0...gatsby-cli@3.0.0-next.1) (2021-02-08)
+
+**Note:** Version bump only for package gatsby-cli
+
+# [3.0.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.20.0-next.1...gatsby-cli@3.0.0-next.0) (2021-02-05)
+
+**Note:** Version bump only for package gatsby-cli
+
+# [2.20.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.20.0-next.0...gatsby-cli@2.20.0-next.1) (2021-02-04)
+
+**Note:** Version bump only for package gatsby-cli
+
+# [2.20.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.19.0-next.1...gatsby-cli@2.20.0-next.0) (2021-01-28)
+
+**Note:** Version bump only for package gatsby-cli
+
+# [2.19.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.19.0-next.0...gatsby-cli@2.19.0-next.1) (2021-01-26)
+
+**Note:** Version bump only for package gatsby-cli
+
 # [2.19.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.18.0-next.1...gatsby-cli@2.19.0-next.0) (2021-01-18)
 
 ### Bug Fixes